refactor: rename TextureDecoder to Texture2DDecoder for consistency
Aligns naming with TextureCubeDecoder.
