+++ packages/loader/src/resource-deserialize/index.ts
@@ -7,7 +7,7 @@ export * from "./resources/animationClip/AnimationClipDecoder";
 export type { IModelMesh } from "./resources/mesh/IModelMesh";
 export { MeshDecoder } from "./resources/mesh/MeshDecoder";
 export { ReflectionParser } from "./resources/parser/ReflectionParser";
-export { Texture2DDecoder } from "./resources/texture2D/TextureDecoder";
+export { Texture2DDecoder } from "./resources/texture2D/Texture2DDecoder";
 export { TextureCubeDecoder } from "./resources/textureCube/TextureCubeDecoder";
 
 /**
